+++ tweets/api/tests.py
@@ -1,0 +1,69 @@
+from rest_framework.test import APIClient
+from testing.testcases import TestCase
+from tweets.models import Tweet
+
+
+# 注意要加 '/' 结尾，要不然会产生 301 redirect
+TWEET_LIST_API = '/api/tweets/'
+TWEET_CREATE_API = '/api/tweets/'
+
+
+class TweetApiTests(TestCase):
+
+    def setUp(self):
+        self.anonymous_client = APIClient()
+
+        self.user1 = self.createUser('user1', 'user1@jiuzhang.com')
+        self.tweets1 = [
+            self.createTweet(self.user1)
+            for i in range(3)
+        ]
+        self.user1_client = APIClient()
+        self.user1_client.force_authenticate(self.user1)
+
+        self.user2 = self.createUser('user2', 'user2@jiuzhang.com')
+        self.tweets2 = [
+            self.createTweet(self.user2)
+            for i in range(2)
+        ]
+
+    def test_list_api(self):
+        # 必须带 user_id
+        response = self.anonymous_client.get(TWEET_LIST_API)
+        self.assertEqual(response.status_code, 400)
+
+        # 正常 request
+        response = self.anonymous_client.get(TWEET_LIST_API, {'user_id': self.user1.id})
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(len(response.data['tweets']), 3)
+        response = self.anonymous_client.get(TWEET_LIST_API, {'user_id': self.user2.id})
+        self.assertEqual(len(response.data['tweets']), 2)
+        # 检测排序是按照新创建的在前面的顺序来的
+        self.assertEqual(response.data['tweets'][0]['id'], self.tweets2[1].id)
+        self.assertEqual(response.data['tweets'][1]['id'], self.tweets2[0].id)
+
+    def test_create_api(self):
+        # 必须登录
+        response = self.anonymous_client.post(TWEET_CREATE_API)
+        self.assertEqual(response.status_code, 403)
+
+        # 必须带 content
+        response = self.user1_client.post(TWEET_CREATE_API)
+        self.assertEqual(response.status_code, 400)
+        # content 不能太短
+        response = self.user1_client.post(TWEET_CREATE_API, {'content': '1'})
+        self.assertEqual(response.status_code, 400)
+        # content 不能太长
+        response = self.user1_client.post(TWEET_CREATE_API, {
+            'content': '0' * 141
+        })
+        self.assertEqual(response.status_code, 400)
+
+        # 正常发帖
+        tweets_count = Tweet.objects.count()
+        response = self.user1_client.post(TWEET_CREATE_API, {
+            'content': 'Hello World, this is my first tweet!'
+        })
+        self.assertEqual(response.status_code, 201)
+        self.assertEqual(response.data['user']['id'], self.user1.id)
+        self.assertEqual(Tweet.objects.count(), tweets_count + 1)
